backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]>=2.0.45
psycopg[binary]>=3.2.13
alembic==1.14.0
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
bcrypt>=4.1
python-multipart==0.0.20
google-api-python-client==2.155.0
google-auth==2.37.0
httpx==0.28.1
openpyxl==3.1.5
